Move server into `web` package
