+++ lib/streamlit/web/server/__init__.py
@@ -11,3 +11,12 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+from .server import (
+    Server as Server,
+    server_address_is_unix_socket as server_address_is_unix_socket,
+)
+
+from .routes import (
+    allow_cross_origin_requests as allow_cross_origin_requests,
+)
